refactor(ui-tests): extract CoralogixUITestCase shared base
The UIKit and SwiftUI UserInteractionUITests held identical isCI detection,
timing constants (elementTimeout/shortDelay/sdkFlushDelay/networkDelay),
setUp/tearDown plumbing, and clearValidationData(). Any timing tweak had to
be applied in both files.

Lift everything into Example/Shared/CoralogixUITestCase.swift, added to
both UI test target memberships. Subclasses now contain only test methods
and target-specific helpers (~84 lines of duplication removed).
